requirements.txt
beautifulsoup4==4.12.3
playwright==1.47.0
langdetect==1.0.9
notion-client==2.2.1
requests==2.32.3
tenacity==9.0.0
certifi==2024.6.2
python-dotenv==1.0.0
